+++ lib/components/custom_app_bar.dart
@@ -1,7 +1,7 @@
 import 'package:final_project_team02/holder/profile/settings_page.dart';
 import 'package:flutter/material.dart';
 
-import '../holder/main_hoder.dart';
+import '../holder/main_holder.dart';
 
 class CustomAppBar extends StatelessWidget implements PreferredSizeWidget {
   final IconData? miconRight;
@@ -20,11 +20,9 @@ class CustomAppBar extends StatelessWidget implements PreferredSizeWidget {
         children: [
           TextButton(
             onPressed: () {
-              Navigator.pushAndRemoveUntil(
+              Navigator.push(
                 context,
                 MaterialPageRoute(builder: (context) => MainHolder()),
-                ModalRoute.withName(
-                    '/home'), // '/firstPage' 라우트를 제외한 모든 페이지를 스택에서 제거
               );
             },
             child: Icon(miconRight),
